+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Fixed
+
+Use correct authorization type for debit payment preauthorization
+
 ## [1.0.1-beta]
 ### Fixed